feat: 构建脚本test

+++ scripts/publish.js
@@ -1,19 +1,32 @@
+/* eslint-disable no-console */
 const shell = require('shelljs');
+const colors = require('colors');
+const argv = require('./argv');
 
-shell.rm('-rf', 'dist');
+const { version, tag = 'latest' } = argv;
 
-shell.cp('-r', 'src', 'dist');
+console.log('tag:', version, tag);
 
-shell.cp('package.json', 'dist');
+if (version) {
+  shell.rm('-rf', 'dist');
 
-shell.cp('README.md', 'dist');
+  shell.cp('-r', 'src', 'dist');
 
-shell.exec('npm config get registry', (_, stdout) => {
-  if (!stdout.includes('registry.npmjs.org')) {
-    shell.exec('npm config set registry=https://registry.npmjs.org');
-  }
-});
+  shell.cp('package.json', 'dist');
 
-shell.exec('npm publish --access=public dist -tag alpha');
+  shell.cp('README.md', 'dist');
 
-shell.exec('curl -X PUT https://npm.taobao.org/sync/eslint-config-kzf');
+  shell.exec('npm config get registry', (_, stdout) => {
+    if (!stdout.includes('registry.npmjs.org')) {
+      shell.exec('npm config set registry=https://registry.npmjs.org');
+    }
+  });
+
+  shell.exec(`npm version ${version}`);
+
+  shell.exec(`npm publish --access=public dist --tag ${tag}`);
+
+  shell.exec('curl -X PUT https://npm.taobao.org/sync/eslint-config-kzf');
+} else {
+  console.error(colors.red('你未指定version 参数'));
+}
